+++ package/index.ts
@@ -82,6 +82,12 @@ export function getUploader(options: IUploadQueueOptions, callbacks: IUploadQueu
     return new Uploader(options, callbacks);
 };
 
+export function getValueOrResult<T>(valueOrGetter?: T | (() => T)): T | undefined {
+    if (typeof valueOrGetter === 'function')
+        return valueOrGetter();
+
+    return valueOrGetter;
+}
 export function newGuid(): string {
     let d = new Date().getTime();
     let uuid = 'xxxxxxxx-xxxx-4xxx-yxxx-xxxxxxxxxxxx'.replace(/[xy]/g, function (c) {
@@ -123,8 +129,8 @@ export interface IOffsetInfo {
 }
 export interface IUploadAreaOptions extends IUploadOptions {
     maxFileSize?: number;
-    allowDragDrop?: boolean;
-    clickable?: boolean;
+    allowDragDrop?: boolean | (() => boolean);
+    clickable?: boolean | (() => boolean);
     accept?: string;
     multiple?: boolean;
     validateExtension?: boolean;
@@ -348,6 +354,7 @@ export class UploadArea {
 
         forEach(this.fileList, (file: IUploadFile) => {
             file.guid = newGuid();
+            delete file.uploadStatus;
             file.url = this.uploadCore.getUrl(file);
             file.onError = this.options.onFileError || (() => { ; });
             file.onCancel = this.options.onFileCanceled || (() => { ; });
@@ -399,27 +406,26 @@ export class UploadArea {
         if (this.formForNoFileApi)
             this.formForNoFileApi.style.display = 'none';
 
-        let onChange = (e: Event) => this.onChange(e);
+        const onChange = (e: Event) => this.onChange(e);
         addEventHandler(this.fileInput, 'change', onChange);
         this.unregisterOnChange = () => removeEventHandler(this.fileInput, 'change', onchange);
 
         if (this.options.multiple) {
             this.fileInput.setAttribute('multiple', '');
         }
-        if (this.options.clickable) {
-            let onClick = () => this.onClick();
-            addEventHandler(this.targetElement, 'click', onClick);
-            this.unregisterOnClick = () => removeEventHandler(this.targetElement, 'click', onClick);
-        }
-        if (this.options.allowDragDrop) {
-            let onDrag = (e: DragEvent) => this.onDrag(e);
-            addEventHandler(this.targetElement, 'dragover', onDrag);
-            this.unregisterOnDragOver = () => removeEventHandler(this.targetElement, 'dragover', onDrag);
 
-            let onDrop = (e: DragEvent) => this.onDrop(e);
-            addEventHandler(this.targetElement, 'drop', onDrop);
-            this.unregisterOnDrop = () => removeEventHandler(this.targetElement, 'drop', onDrop);
-        }
+        const onClick = () => this.onClick();
+        addEventHandler(this.targetElement, 'click', onClick);
+        this.unregisterOnClick = () => removeEventHandler(this.targetElement, 'click', onClick);
+
+        const onDrag = (e: DragEvent) => this.onDrag(e);
+        addEventHandler(this.targetElement, 'dragover', onDrag);
+        this.unregisterOnDragOver = () => removeEventHandler(this.targetElement, 'dragover', onDrag);
+
+        const onDrop = (e: DragEvent) => this.onDrop(e);
+        addEventHandler(this.targetElement, 'drop', onDrop);
+        this.unregisterOnDrop = () => removeEventHandler(this.targetElement, 'drop', onDrop);
+
         // attach to body
         document.body.appendChild(this.fileInput);
     }
@@ -572,6 +578,9 @@ export class UploadArea {
     }
 
     private onDrag(e: DragEvent): void {
+        if (!getValueOrResult(this.options.allowDragDrop))
+            return;
+
         let efct: string | undefined = undefined;
         try {
             efct = e.dataTransfer.effectAllowed;
@@ -581,6 +590,9 @@ export class UploadArea {
     }
 
     private onDrop(e: DragEvent): void {
+        if (!getValueOrResult(this.options.allowDragDrop))
+            return;
+
         this.stopEventPropagation(e);
         if (!e.dataTransfer) {
             return;
@@ -609,6 +621,9 @@ export class UploadArea {
     }
 
     private onClick(): void {
+        if (!getValueOrResult(this.options.clickable))
+            return;
+
         this.fileInput.value = '';
 
         if (this.isIeVersion(10)) {
@@ -947,11 +962,13 @@ export class UploadQueue {
 
     addFiles(files: IUploadFile[]): void {
         forEach(files, file => {
-            this.queuedFiles.push(file);
+            if (!this.queuedFiles.some(queuedFile => queuedFile === file || (!!queuedFile.guid && queuedFile.guid === file.guid))) {
+                this.queuedFiles.push(file);
 
-            file.remove = decorateSimpleFunction(file.remove, () => {
-                this.removeFile(file);
-            });
+                file.remove = decorateSimpleFunction(file.remove, () => {
+                    this.removeFile(file);
+                });
+            }
 
             if (this.callbacks.onFileAddedCallback)
                 this.callbacks.onFileAddedCallback(file);
